README.md
# Assignment-2
This repository is for tp2 assignment 2 test driven development

Assignment2
This is a test driven development program where all of the
Assert functions/methods and some of the test annotations were tested.
Assignment2.2
The program is a basic test driven development program, with test methods 
that complete basic arithmetic equations. The divideValues module is at the top of the
hierarchical program structure, being able to access all the other modules.